[KAFKA] Add test_upgrade.py integration test
