+++ frameworks/kafka/universe/config.json
@@ -49,10 +49,10 @@
             "description":"Kafka Configuration Upgrade parameters",
             "type":"object",
             "properties":{
-              "enable":{
+              "disable":{
                 "description": "Triggers a custom Kafka Configuration Upgrade . Only required if upgrading from an old version, i.e.  dcos-kafka-service version 1.1.19",
                 "type": "boolean",
-                "default": true
+                "default": false
               },
               "cleanup":{
                 "description": "Clean up old TaskInfo and old TaskStatus after Upgrade.",
